refactor modal and word spinner
